refactor!: lift the curse from `get_health` (#115)
the success typing of `Health.get_health/0` was
```
{:ok, %{status: :ready | :bad | :degraded | :error}}
```

this is incredibly cursed, as both `:ok` the `%{status: }` parts are completely
unnecessary.

BREAKING CHANGE: The return type of
`Astarte.DataAccess.Health.Health.get_health/0` changes from `{:ok, %{status:
:ready | :bad | :degraded | :error}}` to `:ready | :bad | :degraded | :error`,
this way users do not need to match on the unnecessary `{:ok, %{status: }}` and
can focus on what matters.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>